Fix pagination by fixing search result caching
Moved logic from models/main to collection/searchresults and
models/searchresult. Code is now cleaner and better readable.
Models/main is a model with queryOptions as it's attributes and has a
searchResults collection which holds all the search results. The
fetching of new results is done in the searchResult model and the caching in the
searchResult collection.