Optimize loading facets
When loading the facets, the browser would hang (yes, hang) when there
where a lot of facets and/or a lot of list options. The chosen solution
is to render the first 50 list options and when scrolling, add 50 when
the user is halfway (load next 50 at 25, load next 50 at 75, ect). On
scroll the scrollbar would change height. This was not a major hiccup,
but by giving the <ul> a predefined height (items * 15px), this is
solved as well. Only (known) bug that exists is when the user scrolls
really fast, s/he will see the empty part of the ul; when the user
scrolls again, the options become visible.