Add results to core lib
The results used to be in a different lib/component, part of
elaborate-modules. The results are now a part of the faceted-
search. This will make it easier and quicker to deploy a
faceted-search instance. The result list can be altered by
providing a result.jade template. The click can be caught by
listening to fs.on('result:click'). If more advanced features
are needed for the result list, one can implement his/her own,
by adding 'results: false' to the initialisation parameters.